+++ test/ae_mdw/db/hardfork_presets_test.exs
@@ -23,9 +23,9 @@ defmodule AeMdw.Db.HardforkPresetsTest do
                  State.new()
                  |> Collection.stream(
                    Model.KindIntTransferTx,
-                   {"accounts_genesis", nil, nil, nil}
+                   {"accounts_roma", nil, nil, nil}
                  )
-                 |> Stream.take_while(&match?({"accounts_genesis", _bi, _target, _txi}, &1))
+                 |> Stream.take_while(&match?({"accounts_roma", _bi, _target, _txi}, &1))
                  |> Enum.count()
 
         assert 325 ==
@@ -79,9 +79,9 @@ defmodule AeMdw.Db.HardforkPresetsTest do
                  State.new()
                  |> Collection.stream(
                    Model.KindIntTransferTx,
-                   {"accounts_genesis", nil, nil, nil}
+                   {"accounts_roma", nil, nil, nil}
                  )
-                 |> Stream.take_while(&match?({"accounts_genesis", _bi, _target, _txi}, &1))
+                 |> Stream.take_while(&match?({"accounts_roma", _bi, _target, _txi}, &1))
                  |> Enum.count()
 
         assert 3 ==
@@ -109,7 +109,7 @@ defmodule AeMdw.Db.HardforkPresetsTest do
 
   describe "mint_sum" do
     test "returns the sum of mintings of a hardfork" do
-      assert HardforkPresets.mint_sum(:genesis) == 89_451_376_822_397_976_634_367_408
+      assert HardforkPresets.mint_sum(:roma) == 89_451_376_822_397_976_634_367_408
       assert HardforkPresets.mint_sum(:minerva) == 27_051_422_546_127_651_538_826_703
       assert HardforkPresets.mint_sum(:fortuna) == 72_681_157_406_723_951_132_840_970
       assert HardforkPresets.mint_sum(:lima) == 86_948_638_040_240_004_719_633_952